+++ src/X/Environment.java
@@ -8,5 +8,4 @@ public class Environment {
     public static Type booleanType, intType, strType, voidType, errorType, floatType, charType;
     public static Function outInt, outFloat, outChar;
 
-
 }