+++ src/X/Nodes/EmptyStructAccessList.java
@@ -1,0 +1,15 @@
+package X.Nodes;
+
+import X.Lexer.Position;
+
+public class EmptyStructAccessList extends List {
+
+    public EmptyStructAccessList(Position pos) {
+        super(pos);
+    }
+
+    public Object visit(Visitor v, Object o) {
+        return v.visitEmptyStructAccessList(this, o);
+    }
+
+}